expand and amend the instr parser.
NOTE: We have to re-evaluate the "RequiredRValOrExpression()" strategy:

There should be a proper expression parser, a recursive one.

It does not need to build the parse tree, only check tokens belong to a whitelist, and
collect identifiers (variables, function invoked).

This needs to be robust, and also tell us the variable names used in the expression.